backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic-settings>=2.0.0
google-genai>=1.0.0
Pillow>=10.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
pytest-mock>=3.14.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
